pyproject.toml
[project]
name = "mPSAT-filter"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "zhengGroupDev", email = "rainyhisy3@163.com"},
]
dependencies = [
    "scipy>=1.10.0",
    "matplotlib>=3.6.3",
    "pyarrow>=11.0.0",
    "tqdm>=4.64.1",
    "einops>=0.6.0",
    "onnx>=1.13.0",
    "onnxruntime-gpu>=1.14.0",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "torchmetrics>=1.4.2",
    "pytorch-lightning>=2.4.0",
    "skl2onnx>=1.17.0",
    "black>=24.8.0",
    "ipykernel>=6.29.5",
    "ultralytics-thop>=2.0.9",
    "scikit-learn==1.2.1",
    "numpy>=1.24.2,<2.0.0",
    "pandas>=1.5.3",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "Apache License 2.0"}


[tool.pdm]
distribution = false
